# docs/api-spec/openapi.yml
openapi: 3.0.3
info:
  title: Authorization API
  description: 認可サーバーのAPI群です。
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: API開発チーム
    email: satoshi.nagashiba@gmail.com
servers:
  - url: https://apis.authorization.com
    description: 本番環境
  - url: https://apis-stg.authorization.com
    description: ステージング環境
  - url: https://apis-dev.authorization.com
    description: 開発環境
  - url: https://apis.authorization-php.dev/api
    description: ローカルPHP（Laravel）環境
  - url: https://apis.authorization-csharp.dev/api
    description: ローカルC#（ASP.NET Core）環境
  - url: https://apis.authorization-go-gin.dev/api
    description: ローカルGo（Gin）環境
  - url: https://apis.authorization-go-beego.dev/api
    description: ローカルGo（Beego）環境
  - url: https://apis.authorization-go-echo.dev/api
    description: ローカルGo（Echo）環境
  - url: https://apis.authorization-java.dev/api
    description: ローカルJava（Spring Boot）環境
  - url: https://apis.authorization-kotlin.dev/api
    description: ローカルKotlin（Ktor）環境
  - url: https://apis.authorization-python.dev/api
    description: ローカルPython（FastAPI）環境
  - url: https://apis.authorization-rb-hanami.dev/api
    description: ローカルRuby（Hanami）環境
  - url: https://apis.authorization-rb-rails.dev/api
    description: ローカルRuby（Rails）環境
  - url: https://apis.authorization-rust.dev/api
    description: ローカルRust（Axum）環境
  - url: https://apis.authorization-ts.dev/api
    description: ローカルTypeScript（Hono）環境
paths:
...
  /clients:
    get:
      operationId: listClients
      summary: クライアント一覧
      tags:
        - clients
      security: []
      parameters:
        - name: keyword
          in: query
          required: false
          description: キーワード
          schema:
            type: string
        - name: start_from
          in: query
          required: false
          description: 利用開始日 From
          schema:
            type: string
            format: date
        - name: start_to
          in: query
          required: false
          description: 利用開始日 To
          schema:
            type: string
            format: date
        - name: statuses
          in: query
          required: false
          description: 状態（複数指定可）
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
        - name: limit
          in: query
          required: false
          description: 取得件数
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: page
          in: query
          required: false
          description: 現在ページ番号（フロントのページャー算出用）
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: sort
          in: query
          required: false
          description: ソート対象カラム
          schema:
            type: string
            enum: [name, status, created_at]
        - name: sort_type
          in: query
          required: false
          description: ソート順
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clients'
        '404':
          description: ユーザーが存在しない場合のエラー
...
  /clients/{identifier}/qr:
    get:
      operationId: getClientQr
      summary: スマホアプリ連携用QRコードデータを取得します
      description: クライアント登録通知メールのリンクを踏んだ際にフロントエンドから呼び出されます。QRコードに埋め込むディープリンクURLを返します。
      tags:
        - clients
      security: []
      parameters:
        - name: identifier
          in: path
          required: true
          schema:
            type: string
          description: クライアント識別子
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientQr'
        '404':
          description: クライアントが存在しない
...
components:
  schemas:
...
    ClientQr:
      type: object
      description: スマホアプリ連携用QRコードデータ
      required:
        - identifier
        - deeplink_url
      properties:
        identifier:
          type: string
          description: クライアント識別子
          example: "6823f4a1c9d2e"
        deeplink_url:
          type: string
          description: QRコードに埋め込むディープリンクURL
          example: "https://app.authorization.com/clients/6823f4a1c9d2e"
...
    Client:
      type: object
      description: クライアント情報を表すオブジェクトです。
      required:
        - id
        - name
        - identifier
        - post_code
        - pref
        - city
        - address
        - tel
        - email
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "株式会社モックデータ商事"
        identifier:
          type: string
          example: "mock"
        post_code:
          type: string
          example: "1234567"
        pref:
          type: string
          example: "東京都"
        city:
          type: string
          example: "架空市中央区みなみ町"
        address:
          type: string
          example: "1丁目2番3号"
        building:
          type: string
          example: "サンプルプラザ東館"
        tel:
          type: string
          example: "0312345678"
        email:
          type: string
          example: "satoshi.nagashiba@gmail.com"
        status:
          type: integer
          example: 1
        start_at:
          type: string
          format: date-time
          example: "2025-11-01T10:00:00Z"
        stop_at:
          type: string
          format: date-time
          example: "2025-11-01T10:00:00Z"
        created_at:
          type: string
          format: date-time
          example: "2025-11-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-11-01T10:00:00Z"
        version:
          type: integer
          format: int64
          description: 楽観排他ロック用バージョン番号
          example: 1
    Clients:
      type: object
      description: クライアント一覧（ページング付き）
      required:
        - data
        - pager
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Client'
        pager:
          $ref: '#/components/schemas/Pager'
...
    Pager:
      type: object
      description: ページャー情報
      required:
        - count
        - limit
        - next
        - previous
        - page
        - nextPage
        - previousPage
        - pageCount
        - first
        - last
        - firstRecordCount
        - lastRecordCount
        - startPage
        - endPage
      properties:
        count:
          type: integer
          description: 総件数
          example: 1000
        limit:
          type: integer
          description: 取得件数
          example: 20
        next:
          type: boolean
          description: 次頁あり
          example: true
        previous:
          type: boolean
          description: 前頁あり
          example: false
        page:
          type: integer
          description: 現在ページ番号
          example: 1
        nextPage:
          type: integer
          description: 次ページ番号
          example: 2
        previousPage:
          type: integer
          description: 前ページ番号
          example: 0
        pageCount:
          type: integer
          description: 総ページ数
          example: 50
        first:
          type: boolean
          description: 先頭ページへの移動可否（page > 1）
          example: false
        last:
          type: boolean
          description: 最終ページへの移動可否（pageCount > page）
          example: true
        firstRecordCount:
          type: integer
          description: 現在ページの先頭レコード番号
          example: 1
        lastRecordCount:
          type: integer
          description: 現在ページの末尾レコード番号
          example: 20
        startPage:
          type: integer
          description: ページャー表示開始ページ
          example: 1
        endPage:
          type: integer
          description: ページャー表示終了ページ
          example: 10
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    clientTokenAuth:
      type: http
      scheme: bearer
      description: クライアントのアクセストークン認証（JWT ではなく登録時に発行される固定トークン）
...